fix: phpstan

+++ src/Services/ResourceConfigGeneratorService.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Adeliom\SyliusEasyCrudPlugin\Services;
 
+use Mockery\Exception;
 use Symfony\Bundle\MakerBundle\Str;
 use Symfony\Bundle\MakerBundle\Util\YamlSourceManipulator;
 use Symfony\Component\Yaml\Yaml;
@@ -14,29 +15,54 @@ class ResourceConfigGeneratorService
 
     public const YAML_RESOURCE_FILE = 'config/packages/sylius_resource.yaml';
 
-    private string $scope;
-
-    public function generateConfig($entityName, $namespaces, $io, $scope = ''): void
-    {
-        $this->scope = $scope;
-
-        $this->appendRoutingConfig($yaml, $entityName);
-        file_put_contents(
-            self::YAML_ROUTES_FILE,
-            Yaml::dump($yaml, 2, 4, Yaml::DUMP_MULTI_LINE_LITERAL_BLOCK),
-            \FILE_APPEND,
-        );
-        $io->writeln('Updated: ' . self::YAML_ROUTES_FILE);
-
-        $yamlGenerator = new YamlSourceManipulator(file_get_contents(self::YAML_RESOURCE_FILE));
-        $yaml = $yamlGenerator->getData();
-        $this->appendResourceConfig($yaml, $entityName, $namespaces);
-        $yamlGenerator->setData($yaml);
-        file_put_contents(self::YAML_RESOURCE_FILE, $yamlGenerator->getContents());
-        $io->writeln('Updated: ' . self::YAML_RESOURCE_FILE);
+    /**
+     * @param array<int|string, string> $namespaces
+     */
+    public function __construct(
+        private string $scope,
+        private array $namespaces,
+    ) {
+        if ($this->scope === '' || 0 === count($this->namespaces)) {
+            throw new Exception('Configurator is not well-declared.');
+        }
     }
 
-    private function appendRoutingConfig(&$data, $entityName): void
+    /**
+     * @return array<int, string>|string
+     */
+    public function generateConfig(string ...$entitiesName): array|string
+    {
+        try {
+            $yaml = [];
+            if (0 === count($entitiesName)) {
+                throw new Exception('At least one entity name must be given.');
+            }
+            foreach ($entitiesName as $entityName) {
+                $this->appendRoutingConfig($yaml, $entityName);
+                file_put_contents(
+                    self::YAML_ROUTES_FILE,
+                    Yaml::dump($yaml, 2, 4, Yaml::DUMP_MULTI_LINE_LITERAL_BLOCK),
+                    \FILE_APPEND,
+                );
+                if (file_exists(self::YAML_RESOURCE_FILE)) {
+                    $yamlGenerator = new YamlSourceManipulator(file_get_contents(self::YAML_RESOURCE_FILE) ?: '');
+                    $yaml = $yamlGenerator->getData();
+                    $this->appendResourceConfig($yaml, $entityName);
+                    $yamlGenerator->setData($yaml);
+                    file_put_contents(self::YAML_RESOURCE_FILE, $yamlGenerator->getContents());
+                }
+            }
+
+            return [self::YAML_ROUTES_FILE, self::YAML_RESOURCE_FILE];
+        } catch (Exception $e) {
+            return $e->getCode() . ' : ' . $e->getMessage();
+        }
+    }
+
+    /**
+     * @param array<string, mixed> $data
+     */
+    private function appendRoutingConfig(array &$data, string $entityName): void
     {
         $resourceDataBlock =
             'alias: happy_cms.' . mb_strtolower($this->scope) . '_' . mb_strtolower(Str::asSnakeCase($entityName)) . "\n"
@@ -79,22 +105,25 @@ class ResourceConfigGeneratorService
             ];
     }
 
-    private function appendResourceConfig(&$data, $entityName, $namespaces): void
+    /**
+     * @param array<string, mixed> $data
+     */
+    private function appendResourceConfig(array &$data, string $entityName): void
     {
         $data['sylius_resource']['resources'][] = YamlSourceManipulator::EMPTY_LINE_PLACEHOLDER_VALUE;
         $data['sylius_resource']['resources']['happy_cms.' . mb_strtolower($this->scope) . '_' . mb_strtolower(Str::asSnakeCase($entityName))] =
             [
                 'driver' => 'doctrine/orm',
                 'classes' => [
-                    'model' => $namespaces['entity'] . $entityName,
-                    'repository' => $namespaces['repository'] . $entityName . 'Repository',
-                    'form' => $namespaces['admin'] . $entityName . 'Admin',
+                    'model' => $this->namespaces['entity'] . $entityName,
+                    'repository' => $this->namespaces['repository'] . $entityName . 'Repository',
+                    'form' => $this->namespaces['admin'] . $entityName . 'Admin',
                 ],
                 'translation' => [
                     'classes' => [
-                        'model' => $namespaces['entity'] . $entityName . 'Translation',
+                        'model' => $this->namespaces['entity'] . $entityName . 'Translation',
                         'controller' => "Adeliom\SyliusEasyCrudPlugin\Controller\SyliusCrudResourceController",
-                        'form' => $namespaces['admin'] . $entityName . 'Admin',
+                        'form' => $this->namespaces['admin'] . $entityName . 'Admin',
                     ],
                 ],
             ];
